requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
pydantic==2.6.1
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
bcrypt==4.1.2
requests==2.31.0 